+++ tests-fixture/src/Cards/WorkHours/Card.php
@@ -3,22 +3,11 @@
 namespace Fixtures\Hello\Cards\WorkHours;
 
 use LaravelUi5\Core\Enums\ArtifactType;
+use LaravelUi5\Core\Ui5\AbstractUi5Card;
 use LaravelUi5\Core\Ui5\Capabilities\DataProviderInterface;
-use LaravelUi5\Core\Ui5\Contracts\Ui5CardInterface;
-use LaravelUi5\Core\Ui5\Contracts\Ui5ModuleInterface;
 
-class Card implements Ui5CardInterface
+class Card extends AbstractUi5Card
 {
-
-    public function __construct(protected Ui5ModuleInterface $module)
-    {
-    }
-
-    public function getModule(): Ui5ModuleInterface
-    {
-        return $this->module;
-    }
-
     public function getType(): ArtifactType
     {
         return ArtifactType::Card;
@@ -31,7 +20,7 @@ class Card implements Ui5CardInterface
 
     public function getNamespace(): string
     {
-        return 'com.laravelui5.hello.work-hours';
+        return 'com.laravelui5.hello.cards.work-hours';
     }
 
     public function getVersion(): string