+++ tests-fixture/src/Hello.php
@@ -8,7 +8,7 @@ class Hello
     public const string NAMESPACE = 'com.laravelui5.hello';
     public const string ROLE = 'Test';
     public const string ACTION_NAME = 'toggleLock';
-    public const string ACTION_SLUG = 'world';
+    public const string ACTION_SLUG = 'com.laravelui5.hello.actions.world';
 
     public static function ui5Config(): array
     {