+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Renamed Integration Tests
 - Added Action Test
 - Register OData routes with const prefix
+- Added `resolvePath` to SourceStrategyInterface
+- Added `CardControllerTest`
 
 ## [2.4.9] - 2026-01-17
 - Fetch XSRF-Token early